pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "bidsmosaic"
version = "0.0.1"
dependencies = [
    "numpy",
    "pybids",
    "nilearn",
    "fpdf2",
    "pillow"
]

[project.scripts]
bids-mosaic = "bidsmosaic.mosaic:main"
